rename to `_options.py`
