+++ pypsa/_options.py
